Reorganização na estrutura da pasta de components.
